Currency_app/requirements.txt
altgraph==0.17.3
appdirs==1.4.4
asgiref==3.6.0
asttokens==2.0.5
audioread==2.1.9
backcall==0.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.4
cryptography==39.0.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.1.6
djangorestframework==3.14.0
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
fonttools==4.33.3
future==0.18.2
greenlet==1.1.2
httpie==3.2.1
idna==3.3
importlib-metadata==4.12.0
inflect==5.3.0
ipykernel==6.10.0
ipython==8.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpatch==1.32
jsonpointer==2.3
jupyter-client==7.2.0
jupyter-core==4.9.2
kiwisolver==1.4.2
librosa==0.8.1
llvmlite==0.38.1
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
multidict==6.0.4
mysql-connector==2.2.9
nest-asyncio==1.5.4
numba==0.55.2
numpy==1.20.3
packaging==21.3
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.2
pooch==1.6.0
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
pycparser==2.21
pycryptodome==3.17
pygame==2.1.2
Pygments==2.14.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pynndescent==0.5.7
pyparsing==3.0.9
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqt5-tools==5.15.4.3.2
PyQt5Designer==5.14.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
pywin32==303
pywin32-ctypes==0.2.0
pyzmq==22.3.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
requests==2.27.1
requests-toolbelt==0.10.1
resampy==0.2.2
rich==13.3.2
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sounddevice==0.4.3
SoundFile==0.10.3.post1
spyder-kernels==2.2.0
SQLAlchemy==1.4.39
sqlparse==0.4.3
stack-data==0.2.0
threadpoolctl==3.1.0
torchfile==0.1.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
tzdata==2022.7
umap-learn==0.5.2
Unidecode==1.3.2
urllib3==1.26.7
virtualenv==20.14.1
visdom==0.1.8.9
wcwidth==0.2.5
webrtcvad==2.0.10
websocket-client==1.3.2
Werkzeug==2.2.1
zipp==3.8.1
